==> Python/00_exploring energy poverty monitor_19_22.ipynb <==
import pandas as pd 
import os
import geopandas as gpd 
import zipfile 
import matplotlib.pyplot as plt
import folium
import pickle 
import seaborn as sns
sns.set_theme()
import numpy as np 
import pyogrio
pyogrio.list_drivers()  
import folium

os.chdir("C:/Users/tejed002/OneDrive - Wageningen University & Research/PhD-WUR/01_WUR-Research/05_Papers/02_Paper_energy_poverty/Data/Raw/")
os.listdir()

#energy_poverty = pd.read_excel("Energy Poverty Monitor/MonitorEnergiearmoede2019-2022_def.xlsx", sheet_name='Tabel 3', skiprows=3, nrows = 351175)
buurt_2019_2022 = pd.read_csv("ShapeFiles/buurt_homologation_2019_2022.csv", sep = "|", header=0)

# dict_energy_poverty = {"energy_poverty": energy_poverty}
# with open("Energy Poverty Monitor/energy_poverty.pkl", "wb") as file:
#     pickle.dump(dict_energy_poverty, file)
with open("Energy Poverty Monitor/energy_poverty.pkl", 'rb') as file:
    dict_energy_poverty = pickle.load(file)
energy_poverty = dict_energy_poverty["energy_poverty"]

fua = gpd.read_file("ShapeFiles/fuas/fuas.shp")
fua = fua[fua.iso3 == "NLD"]

with open('ShapeFiles/00_Processed/dict_gdf.pkl', 'rb') as file:
    dict_gdf = pickle.load(file)

with open('ShapeFiles/00_Processed/dict_wijk_gdf.pkl', 'rb') as file:
    dict_wijk_gdf = pickle.load(file)

dict_wijk_gdf["2022"].shape

with open('ShapeFiles/00_Processed/dict_gemeente_gdf.pkl', 'rb') as file:
    dict_gemeente_gdf = pickle.load(file)

years = [2019, 2020, 2021, 2022, 2023]
for year in years:
    gdf_1 = dict_gdf[str(year)]
    gdf_1 = gdf_1[gdf_1["H2O"] == "NEE"]
    dict_gdf[str(year)] = gdf_1
    
    gdf_2 = dict_wijk_gdf[str(year)]
    gdf_2 = gdf_2[gdf_2["H2O"] == "NEE"]
    dict_wijk_gdf[str(year)] = gdf_2
    
    gdf_3 = dict_gemeente_gdf[str(year)]
    gdf_3 = gdf_3[gdf_3["H2O"] == "NEE"]
    dict_gemeente_gdf[str(year)] = gdf_3

# columns_buurten = ["BU_CODE", "BU_NAAM", "WK_CODE", "GM_CODE", "GM_NAAM", "WK_CODE", "GM_CODE", "GM_NAAM", "IND_WBI", 'JRSTATCODE', 'JAAR', 'geometry']

gdf_buurt = [f"gdf_buurt_{ii}" for ii in years]
gdf_wijk = [f"gdf_wijk_{ii}" for ii in years]
gdf_gemeente = [f"gdf_gemeente_{ii}" for ii in years]

for name, value in zip(gdf_buurt, dict_gdf.values()):
    globals()[name] = value

for name, value in zip(gdf_wijk, dict_wijk_gdf.values()):
    globals()[name] = value

for name, value in zip(gdf_gemeente, dict_gemeente_gdf.values()):
    globals()[name] = value

column_names = {
    'Soort eigendom woning': 'property_type',
    'Totaal huishoudens1 (aantal)': 'total_households',
    'Energiearmoede volgens LIHE en/of LILEK2 (%)': 'LIHE_AndOr_LILEK',
    'Energiearmoede volgens LIHE en/of LIZLEK2 (%)': 'LIHE_AndOr_LIZLEK',
    'Energiearmoede volgens HEq2 (%)': 'HEQ',
    'Energiearmoede volgens LIHE2 (%)': 'LIHE',
    'Energiearmoede volgens LILEK2 (%)': 'LILEK',
    'Energiearmoede volgens LIZLEK2 (%)': 'LIZLEK ',
    'Energiearmoede volgens LEKWI2 (%)': 'LEKWI',
    'Energiearmoede volgens ZLEKWI2 (%)': 'ZLEKWI',
    'Laag inkomen, LI2 (%)': 'LI',
    'Hoge energiekosten, HE2 (%)': 'HE',
    'Lage energetische kwaliteit, LEK2 (%)': 'LEK',
    'Zeer lage energetische kwaliteit, ZLEK2 (%)': 'ZLEK',
    'Gestandaardiseerd residueel inkomen \n(x 1 000 euro)': 'standardize_residual_income',
    'Gemiddelde energiequote (%)': 'average_energy_ratio',
    "Jaar" : "year",
    }

property_type = {
    'Totaal':"Totaal", 
     'Koop':"Owner", # Buy 
     'Huur (corporatie)': "Rental_Corporation", # rent(other)
     'Huur (overig)': "Rental_other", # rent(other) 
     'Onbekend': "Unknown" # "unknown"
     } 

energy_poverty = energy_poverty.rename(columns = column_names)
energy_poverty['property_type'] = energy_poverty['property_type'].replace(property_type)
energy_poverty.head()

# Total energy poverty

energy_poverty_22 = energy_poverty[energy_poverty['year'] == 2022]
energy_poverty_21 = energy_poverty[energy_poverty['year'] == 2021]
energy_poverty_20 = energy_poverty[energy_poverty['year'] == 2020]
energy_poverty_19 = energy_poverty[energy_poverty['year'] == 2019]


ep_geemente_19 = energy_poverty_19[(energy_poverty_19["property_type"] == "Totaal") & (energy_poverty_19["Buurt"] == "Totaal") & (energy_poverty_19["Wijk"] == "Totaal") & (energy_poverty_19["Gemeente"] != "Totaal")]
ep_geemente_20 = energy_poverty_20[(energy_poverty_20["property_type"] == "Totaal") & (energy_poverty_20["Buurt"] == "Totaal") & (energy_poverty_20["Wijk"] == "Totaal") & (energy_poverty_20["Gemeente"] != "Totaal")]
ep_geemente_21 = energy_poverty_21[(energy_poverty_21["property_type"] == "Totaal") & (energy_poverty_21["Buurt"] == "Totaal") & (energy_poverty_21["Wijk"] == "Totaal") & (energy_poverty_21["Gemeente"] != "Totaal")]
ep_geemente_22 = energy_poverty_22[(energy_poverty_22["property_type"] == "Totaal") & (energy_poverty_22["Buurt"] == "Totaal") & (energy_poverty_22["Wijk"] == "Totaal") & (energy_poverty_22["Gemeente"] != "Totaal")]


ep_buurten_19 = energy_poverty_19[(energy_poverty_19["property_type"] == "Totaal") & (energy_poverty_19["Buurt"] != "Totaal")]
ep_buurten_20 = energy_poverty_20[(energy_poverty_20["property_type"] == "Totaal") & (energy_poverty_20["Buurt"] != "Totaal")]
ep_buurten_21 = energy_poverty_21[(energy_poverty_21["property_type"] == "Totaal") & (energy_poverty_21["Buurt"] != "Totaal")]
ep_buurten_22 = energy_poverty_22[(energy_poverty_22["property_type"] == "Totaal") & (energy_poverty_22["Buurt"] != "Totaal")]

print(sum(~ep_buurten_19.LIHE_AndOr_LILEK.isna())/len(ep_buurten_19.LIHE_AndOr_LILEK))
print(sum(~ep_buurten_20.LIHE_AndOr_LILEK.isna())/len(ep_buurten_20.LIHE_AndOr_LILEK))
print(sum(~ep_buurten_21.LIHE_AndOr_LILEK.isna())/len(ep_buurten_21.LIHE_AndOr_LILEK))
print(sum(~ep_buurten_22.LIHE_AndOr_LILEK.isna())/len(ep_buurten_22.LIHE_AndOr_LILEK))


os.getcwd()

ep_wijken_19 = energy_poverty_19[(energy_poverty_19["property_type"] == "Totaal") & (energy_poverty_19["Buurt"] == "Totaal") & (energy_poverty_19["Wijk"] != "Totaal")]
ep_wijken_20 = energy_poverty_20[(energy_poverty_20["property_type"] == "Totaal") & (energy_poverty_20["Buurt"] == "Totaal") & (energy_poverty_20["Wijk"] != "Totaal")]
ep_wijken_21 = energy_poverty_21[(energy_poverty_21["property_type"] == "Totaal") & (energy_poverty_21["Buurt"] == "Totaal") & (energy_poverty_21["Wijk"] != "Totaal")]
ep_wijken_22 = energy_poverty_22[(energy_poverty_22["property_type"] == "Totaal") & (energy_poverty_22["Buurt"] == "Totaal") & (energy_poverty_22["Wijk"] != "Totaal")]

dict_ep_wijken = {
    "19": ep_wijken_19,
    "20": ep_wijken_20,
    "21": ep_wijken_21,
    "22": ep_wijken_22,
}
# C:\Users\tejed002\OneDrive - Wageningen University & Research\PhD-WUR\01_WUR-Research\05_Papers\02_Paper_energy_poverty\analysis\data\wijken\energy_poverty
with open("../../analysis/data/wijken/energy_poverty/dict_ep_wijken.pkl", "wb") as file:
    pickle.dump(dict_ep_wijken, file)

# Rental and owner 

energy_poverty_19["property_type"].unique()
list_property = ['Owner', 'Rental_Corporation', 'Rental_other']

for property in list_property:
    ep_geemente_19 = energy_poverty_19[(energy_poverty_19["property_type"] == property) & (energy_poverty_19["Buurt"] == "Totaal") & (energy_poverty_19["Wijk"] == "Totaal") & (energy_poverty_19["Gemeente"] != "Totaal")]
    ep_geemente_20 = energy_poverty_20[(energy_poverty_20["property_type"] == property) & (energy_poverty_20["Buurt"] == "Totaal") & (energy_poverty_20["Wijk"] == "Totaal") & (energy_poverty_20["Gemeente"] != "Totaal")]
    ep_geemente_21 = energy_poverty_21[(energy_poverty_21["property_type"] == property) & (energy_poverty_21["Buurt"] == "Totaal") & (energy_poverty_21["Wijk"] == "Totaal") & (energy_poverty_21["Gemeente"] != "Totaal")]
    ep_geemente_22 = energy_poverty_22[(energy_poverty_22["property_type"] == property) & (energy_poverty_22["Buurt"] == "Totaal") & (energy_poverty_22["Wijk"] == "Totaal") & (energy_poverty_22["Gemeente"] != "Totaal")]
    
    ep_buurten_19 = energy_poverty_19[(energy_poverty_19["property_type"] == property) & (energy_poverty_19["Buurt"] != "Totaal")]
    ep_buurten_20 = energy_poverty_20[(energy_poverty_20["property_type"] == property) & (energy_poverty_20["Buurt"] != "Totaal")]
    ep_buurten_21 = energy_poverty_21[(energy_poverty_21["property_type"] == property) & (energy_poverty_21["Buurt"] != "Totaal")]
    ep_buurten_22 = energy_poverty_22[(energy_poverty_22["property_type"] == property) & (energy_poverty_22["Buurt"] != "Totaal")]
    
    ep_wijken_19 = energy_poverty_19[(energy_poverty_19["property_type"] == property) & (energy_poverty_19["Buurt"] == "Totaal") & (energy_poverty_19["Wijk"] != "Totaal")]
    ep_wijken_20 = energy_poverty_20[(energy_poverty_20["property_type"] == property) & (energy_poverty_20["Buurt"] == "Totaal") & (energy_poverty_20["Wijk"] != "Totaal")]
    ep_wijken_21 = energy_poverty_21[(energy_poverty_21["property_type"] == property) & (energy_poverty_21["Buurt"] == "Totaal") & (energy_poverty_21["Wijk"] != "Totaal")]
    ep_wijken_22 = energy_poverty_22[(energy_poverty_22["property_type"] == property) & (energy_poverty_22["Buurt"] == "Totaal") & (energy_poverty_22["Wijk"] != "Totaal")]
    
    dict_ep_geemente = {
        "19": ep_geemente_19,
        "20": ep_geemente_20,
        "21": ep_geemente_21,
        "22": ep_geemente_22,
    }
    dict_ep_buurt = {
        "19": ep_buurten_19,
        "20": ep_buurten_20,
        "21": ep_buurten_21,
        "22": ep_buurten_22,
    }
    dict_ep_wijken = {
        "19": ep_wijken_19,
        "20": ep_wijken_20,
        "21": ep_wijken_21,
        "22": ep_wijken_22,
    }
    # C:\Users\tejed002\OneDrive - Wageningen University & Research\PhD-WUR\01_WUR-Research\05_Papers\02_Paper_energy_poverty\analysis\data\wijken\energy_poverty
    path_export_gm = "../../analysis/data/wijken/energy_poverty/" + "dict_ep_gemeente_" + property + ".pkl"
    path_export_bu = "../../analysis/data/wijken/energy_poverty/" + "dict_ep_buurt_" + property + ".pkl"
    path_export_wi = "../../analysis/data/wijken/energy_poverty/" + "dict_ep_wijken_" + property + ".pkl"
    print(path_export_gm)
    with open(path_export_gm, "wb") as file:
        pickle.dump(dict_ep_geemente, file)
        
    with open(path_export_bu, "wb") as file:
        pickle.dump(dict_ep_buurt, file)
    
    with open(path_export_wi, "wb") as file:
        pickle.dump(dict_ep_wijken, file)

ep_geemente_19 = energy_poverty_19[(energy_poverty_19["property_type"] == "Totaal") & (energy_poverty_19["Buurt"] == "Totaal") & (energy_poverty_19["Wijk"] == "Totaal") & (energy_poverty_19["Gemeente"] != "Totaal")]
ep_geemente_20 = energy_poverty_20[(energy_poverty_20["property_type"] == "Totaal") & (energy_poverty_20["Buurt"] == "Totaal") & (energy_poverty_20["Wijk"] == "Totaal") & (energy_poverty_20["Gemeente"] != "Totaal")]
ep_geemente_21 = energy_poverty_21[(energy_poverty_21["property_type"] == "Totaal") & (energy_poverty_21["Buurt"] == "Totaal") & (energy_poverty_21["Wijk"] == "Totaal") & (energy_poverty_21["Gemeente"] != "Totaal")]
ep_geemente_22 = energy_poverty_22[(energy_poverty_22["property_type"] == "Totaal") & (energy_poverty_22["Buurt"] == "Totaal") & (energy_poverty_22["Wijk"] == "Totaal") & (energy_poverty_22["Gemeente"] != "Totaal")]


ep_buurten_19 = energy_poverty_19[(energy_poverty_19["property_type"] == "Totaal") & (energy_poverty_19["Buurt"] != "Totaal")]
ep_buurten_20 = energy_poverty_20[(energy_poverty_20["property_type"] == "Totaal") & (energy_poverty_20["Buurt"] != "Totaal")]
ep_buurten_21 = energy_poverty_21[(energy_poverty_21["property_type"] == "Totaal") & (energy_poverty_21["Buurt"] != "Totaal")]
ep_buurten_22 = energy_poverty_22[(energy_poverty_22["property_type"] == "Totaal") & (energy_poverty_22["Buurt"] != "Totaal")]

print(sum(~ep_buurten_19.LIHE_AndOr_LILEK.isna())/len(ep_buurten_19.LIHE_AndOr_LILEK))
print(sum(~ep_buurten_20.LIHE_AndOr_LILEK.isna())/len(ep_buurten_20.LIHE_AndOr_LILEK))
print(sum(~ep_buurten_21.LIHE_AndOr_LILEK.isna())/len(ep_buurten_21.LIHE_AndOr_LILEK))
print(sum(~ep_buurten_22.LIHE_AndOr_LILEK.isna())/len(ep_buurten_22.LIHE_AndOr_LILEK))


os.getcwd()

ep_wijken_19 = energy_poverty_19[(energy_poverty_19["property_type"] == "Totaal") & (energy_poverty_19["Buurt"] == "Totaal") & (energy_poverty_19["Wijk"] != "Totaal")]
ep_wijken_20 = energy_poverty_20[(energy_poverty_20["property_type"] == "Totaal") & (energy_poverty_20["Buurt"] == "Totaal") & (energy_poverty_20["Wijk"] != "Totaal")]
ep_wijken_21 = energy_poverty_21[(energy_poverty_21["property_type"] == "Totaal") & (energy_poverty_21["Buurt"] == "Totaal") & (energy_poverty_21["Wijk"] != "Totaal")]
ep_wijken_22 = energy_poverty_22[(energy_poverty_22["property_type"] == "Totaal") & (energy_poverty_22["Buurt"] == "Totaal") & (energy_poverty_22["Wijk"] != "Totaal")]

dict_ep_wijken = {
    "19": ep_wijken_19,
    "20": ep_wijken_20,
    "21": ep_wijken_21,
    "22": ep_wijken_22,
}
# C:\Users\tejed002\OneDrive - Wageningen University & Research\PhD-WUR\01_WUR-Research\05_Papers\02_Paper_energy_poverty\analysis\data\wijken\energy_poverty
with open("../../analysis/data/wijken/energy_poverty/dict_ep_wijken.pkl", "wb") as file:
    pickle.dump(dict_ep_wijken, file)

# Processing

ep_wijken = pd.concat([ep_wijken_19, ep_wijken_20, ep_wijken_21, ep_wijken_22])

####### Exploration LIHE, HEQ, and LIHE_AndOr_LILEK foor all the years
ep_wijken_dict = {
    19: ep_wijken_19,
    20: ep_wijken_20,
    21: ep_wijken_21,
    22: ep_wijken_22
}

# Loop through each DataFrame by year and calculate the proportion of non-NaN values in LIHE
for indicator in ["LIHE", "HEQ", "LIHE_AndOr_LILEK"]:
    print(f"-------{indicator}-----------------")
    for year, df in ep_wijken_dict.items():
        proportion_non_nan = sum(~df[indicator].isna()) / len(df[indicator])
        print(f"Year 20{year}: {proportion_non_nan:.2f}")

print(ep_wijken_22.Regiocode.nunique())
print(ep_wijken_19.Regiocode.nunique())
print(len(ep_wijken_22.Wijk))
ep_wijken_22.Wijk[ep_wijken_22.Wijk.duplicated()]
ep_wijken_22[ep_wijken_22.Wijk.isin(["Wijk 02 Zuid"])]

print(gdf_buurt_2023.shape)
gdf_buurt_2023 = gdf_buurt_2023.merge(ep_buurten_22, left_on='BU_CODE', right_on='Regiocode', how = 'left')
print(gdf_buurt_2023.shape)

print(gdf_wijk_2023.shape)
gdf_wijk_2023 = gdf_wijk_2023.merge(ep_wijken_22, left_on='WK_CODE', right_on='Regiocode', how = 'left')
print(gdf_wijk_2023.shape)

print(gdf_wijk_2022.shape)
gdf_wijk_2022 = gdf_wijk_2022.merge(ep_wijken_22, left_on='WK_CODE', right_on='Regiocode', how = 'left')
print(gdf_wijk_2022.shape)

#gdf_gemeente_19_22.to_file('gdf_gemeente_19_22.gpkg', driver='GPKG')
gdf_wijk_2023.to_file('../../analysis/data/wijken/gdf_wijk_2023.gpkg', driver='GPKG')
gdf_wijk_2023.to_file('../../analysis/data/wijken/gdf_wijk_2023.shp')
gdf_wijk_2022.to_file('../../analysis/data/wijken/gdf_wijk_2022.shp')

gdf_wijk_2023.LIHE.describe()
print(sum(gdf_wijk_2023.LIHE.isna())/len(gdf_wijk_2023.LIHE))
print(sum(gdf_wijk_2023.HEQ.isna())/len(gdf_wijk_2023.HEQ))
print(sum(gdf_wijk_2023.LIHE_AndOr_LILEK.isna())/len(gdf_wijk_2023.LIHE_AndOr_LILEK))

year_2019 = ep_geemente_19[['Regiocode', 'LIHE_AndOr_LILEK', 'total_households']]
year_2020 = ep_geemente_20[['Regiocode', 'LIHE_AndOr_LILEK', 'total_households']]
year_2021 = ep_geemente_21[['Regiocode', 'LIHE_AndOr_LILEK', 'total_households']]
year_2022 = ep_geemente_22[['Regiocode', 'LIHE_AndOr_LILEK', 'total_households']]     

year_2019 = year_2019.add_suffix('_2019') 
year_2020 = year_2020.add_suffix('_2020') 
year_2021 = year_2021.add_suffix('_2021') 
year_2022 = year_2022.add_suffix('_2022')          

print(year_2019.LIHE_AndOr_LILEK_2019.describe())
print(year_2020.LIHE_AndOr_LILEK_2020.describe())

print(year_2021.LIHE_AndOr_LILEK_2021.describe())
print(year_2022.LIHE_AndOr_LILEK_2022.describe())

print(gdf_gemeente_2023.shape)
gdf_gemeente_19_22 = gdf_gemeente_2023.merge(year_2019, left_on='GM_CODE', right_on='Regiocode_2019', how = 'left')
gdf_gemeente_19_22 = gdf_gemeente_19_22.merge(year_2020, left_on='GM_CODE', right_on='Regiocode_2020', how = 'left')
gdf_gemeente_19_22 = gdf_gemeente_19_22.merge(year_2021, left_on='GM_CODE', right_on='Regiocode_2021', how = 'left')
gdf_gemeente_19_22 = gdf_gemeente_19_22.merge(year_2022, left_on='GM_CODE', right_on='Regiocode_2022', how = 'left')
print(gdf_gemeente_19_22.shape)

gdf_gemeente_2023.AANT_INW.sum()

#gdf_gemeente_19_22.to_file('gdf_gemeente_19_22.gpkg', driver='GPKG')
gdf_wijk_2023.to_file('gdf_wijk_2023.gpkg', driver='GPKG')


fua = fua.to_crs(gdf_wijk_2023.crs)

gdf_wijk_2023['id'] = gdf_wijk_2023.index.astype(str) 
gdf_wijk_2023['id']

if gdf_wijk_2023.crs != "EPSG:3857":
    gdf_wijk_2023 = gdf_wijk_2023.to_crs(epsg=3857)
print(gdf_wijk_2023.crs)

# Add the GeoDataFrame as a GeoJSON layer to the folium map with transparency
# Define your style function for transparency (optional, only needed if adding GeoJson)
style_function = lambda x: {
    'fillColor': '#ff7800',  # Default fill color (will be overridden by Choropleth)
    'color': 'black',        # Boundary color
    'weight': 1,             # Boundary width
    'fillOpacity': 0.5       # Transparency for GeoJson layer
}

m = folium.Map(location=[52.1326, 5.2913], zoom_start=8)


folium.Choropleth(
    geo_data=gdf_wijk_2023.to_json(),        # Convert GeoDataFrame to GeoJSON format
    name='choropleth',
    data=gdf_wijk_2023,                      # GeoDataFrame as data source
    columns=['id', 'LIHE_AndOr_LILEK'],           # Columns to use (replace 'id' with the column uniquely identifying each geometry)
    key_on='feature.properties.id',# Key to match in GeoJSON (adjust if your unique identifier column is named differently)
    fill_color='YlOrRd',           # Color scale (e.g., 'YlOrRd', 'BuPu', 'RdYlGn')
    fill_opacity=0.7,              # Transparency of the fill
    line_opacity=0.2,              # Transparency of the borders
    legend_name='Column W Values'  # Legend title
).add_to(m)

folium.GeoJson(gdf_wijk_2023.to_json(), style_function=style_function).add_to(m)
folium.LayerControl().add_to(m)
m

# Quick plot
# ax = gdf_wijk_2023[gdf_wijk_2023["GM_NAAM"] == 'Wageningen'].plot(figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color = "gray")
# gdf_wijk_2023[gdf_wijk_2023["GM_NAAM"] == 'Wageningen'].plot(ax = ax, column = 'LIHE_AndOr_LILEK', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="plasma")

ax = gdf_wijk_2023.plot(figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color = "gray")
gdf_wijk_2023.plot(ax = ax, column = 'LIHE', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="plasma")
# Set the title and display the plot
plt.title(f"Plot of NL - Buurten", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")
plt.grid(True)
plt.show()

# Quick plot
ax = gdf_buurt_2023.plot(figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color = "gray")
fua.plot(ax = ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color = "blue")
#ax = gdf_20[gdf_20["BU_NAAM"] == "Verspreide huizen Sleen"].plot(figsize=(10, 10), edgecolor='k', linewidth=2.5)
# gdf_22[gdf_22["BU_CODE"] == "BU01145109"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'red')
#gdf_22[gdf_22["BU_NAAM"] == "Verspreide huizen Sleen"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'green')

# Set the title and display the plot
plt.title(f"Plot of NL - Buurten", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")
plt.grid(True)
plt.show()

# Quick plot
ax = gdf_buurt_2023.plot(figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color = "gray")
gdf_buurt_2023.plot(ax = ax, column = 'LIHE_AndOr_LILEK', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="plasma")
#ax = gdf_20[gdf_20["BU_NAAM"] == "Verspreide huizen Sleen"].plot(figsize=(10, 10), edgecolor='k', linewidth=2.5)
# gdf_22[gdf_22["BU_CODE"] == "BU01145109"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'red')
#gdf_22[gdf_22["BU_NAAM"] == "Verspreide huizen Sleen"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'green')

# Set the title and display the plot
plt.title(f"Plot of NL - Buurten", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")
plt.grid(True)
plt.show()

# Quick plot
ax = gdf_wijk_2023.plot(figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color = "gray")
gdf_wijk_2023.plot(ax = ax, column = 'LIHE_AndOr_LILEK', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="plasma")
#ax = gdf_20[gdf_20["BU_NAAM"] == "Verspreide huizen Sleen"].plot(figsize=(10, 10), edgecolor='k', linewidth=2.5)
# gdf_22[gdf_22["BU_CODE"] == "BU01145109"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'red')
#gdf_22[gdf_22["BU_NAAM"] == "Verspreide huizen Sleen"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'green')

# Set the title and display the plot
plt.title(f"Plot of NL", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")
plt.grid(True)
plt.show()

### which geemente have more missing wijken missing 
missing_geemente = gdf_wijk_2023.groupby(["GM_CODE", "GM_NAAM"])
missing_geemente = missing_geemente["LIHE_AndOr_LILEK"].agg(NA_count = lambda x: x.isna().sum(), 
                                                   NA_perc=lambda x: (x.isna().sum() / len(x) * 100),
                                                   Total = lambda x: len(x)).reset_index()
missing_geemente

print(gdf_gemeente_2023.shape)
gdf_gemeente_2023 = gdf_gemeente_2023.merge(missing_geemente, left_on='GM_CODE', right_on='GM_CODE', how = 'left')
print(gdf_gemeente_2023.shape)

# Quick plot
ax = gdf_gemeente_2023[gdf_gemeente_2023["NA_perc"]<10].plot(figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color = "blue")
gdf_gemeente_2023[gdf_gemeente_2023["NA_perc"]>10].plot(ax = ax, column = 'NA_perc', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="Reds")
#ax = gdf_20[gdf_20["BU_NAAM"] == "Verspreide huizen Sleen"].plot(figsize=(10, 10), edgecolor='k', linewidth=2.5)
# gdf_22[gdf_22["BU_CODE"] == "BU01145109"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'red')
#gdf_22[gdf_22["BU_NAAM"] == "Verspreide huizen Sleen"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'green')

# Set the title and display the plot
plt.title(f"Plot of NL", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")
plt.grid(True)
plt.show()

# plotting histogram for carat using distplot()
print(gdf_gemeente_2023.NA_perc.describe())
sns.histplot(gdf_gemeente_2023.NA_perc, kde=True)

print(energy_poverty_22.Regiocode.isin(gdf_buurt_2022.BU_CODE).value_counts())
print(gdf_buurt_2023.BU_CODE.isin(energy_poverty_22.Regiocode).value_counts())
print(gdf_buurt_2022.BU_NAAM.isin(energy_poverty_22.Buurt).value_counts())


#### Checking in the FUA missing values 
gdf_buurt_2023_fua = gpd.sjoin(gdf_buurt_2023, fua, predicate='within')
gdf_wijk_2023_fua_within = gpd.sjoin(gdf_wijk_2023, fua,  predicate='within')
gdf_gemeente_2023_fua = gpd.sjoin(gdf_gemeente_2023, fua, how='left', predicate='intersects')
fua_buurt_2023 = gpd.sjoin(fua, gdf_buurt_2023, predicate='intersects')

gdf_wijk_2023_fua.geometry.intersection

# Perform the spatial join with the 'intersects' predicate
gdf_wijk_2023_fua = gpd.sjoin(gdf_wijk_2023, fua, predicate='intersects')
# Merge the original 'fua' geometries back into the result to align geometries
gdf_wijk_2023_fua = gdf_wijk_2023_fua.merge(fua[['geometry']], left_on='index_right', right_index=True, suffixes=('', '_fua'))

# Now calculate the intersection
gdf_wijk_2023_fua['intersection'] = gdf_wijk_2023_fua.geometry.intersection(gdf_wijk_2023_fua['geometry_fua'])

# # # Calculate the area of the original geometries in gdf_wijk_2023
gdf_wijk_2023_fua['area_wijk'] = gdf_wijk_2023_fua.geometry.area

# # # Calculate the area of the intersection geometries
gdf_wijk_2023_fua['intersection_area'] = gdf_wijk_2023_fua['intersection'].area

# # Calculate percentage of overlap (intersection area / original area)
gdf_wijk_2023_fua['overlap_percentage'] = (gdf_wijk_2023_fua['intersection_area'] / gdf_wijk_2023_fua['area_wijk']) * 100

# # Filter the spatial join to keep only those geometries that are 'within' a certain overlap percentage
# # For example, keeping only those with at least 70% overlap
threshold = 90  # Set your desired percentage threshold here
gdf_wijk_2023_fua_filtered = gdf_wijk_2023_fua[gdf_wijk_2023_fua['overlap_percentage'] >= threshold]

gdf_wijk_2023_fua['overlap_percentage'].describe().round()

# Quick plot
ax = gdf_wijk_2023.plot(figsize=(10, 10), edgecolor='k', linewidth=0.01, legend=True, color = "white")
# gdf_gemeente_2023_fua[~gdf_gemeente_2023_fua.fuaname_en.isna()].plot(ax = ax, column = 'NA_perc', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="Reds")
#fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red", alpha = 0.3)
gdf_wijk_2023_fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red", alpha = 0.3)
gdf_wijk_2023_fua_filtered.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="blue", alpha = 0.3)
#gdf_buurt_2023_fua[~gdf_buurt_2023_fua.fuaname_en.isna()].plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red")
#ax = gdf_20[gdf_20["BU_NAAM"] == "Verspreide huizen Sleen"].plot(figsize=(10, 10), edgecolor='k', linewidth=2.5)
# gdf_22[gdf_22["BU_CODE"] == "BU01145109"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'red')
#gdf_22[gdf_22["BU_NAAM"] == "Verspreide huizen Sleen"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'green')

# Set the title and display the plot
plt.title(f"Plot of NL", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")

# plt.xlim([min_longitude, max_longitude])  # Replace with your desired x-range (longitude)
# plt.ylim([min_latitude, max_latitude])    # Replace with your desired y-range (latitude)

plt.grid(True)
plt.show()

# Quick plot
ax = gdf_wijk_2023.plot(figsize=(10, 10), edgecolor='k', linewidth=0.01, legend=True, color = "white")
# gdf_gemeente_2023_fua[~gdf_gemeente_2023_fua.fuaname_en.isna()].plot(ax = ax, column = 'NA_perc', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="Reds")
#fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red", alpha = 0.3)
gdf_wijk_2023[gdf_wijk_2023["STED"].isin([1,2,3])].plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.001, legend=True, color="blue", alpha = 0.3)
gdf_wijk_2023_fua_filtered.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.001, legend=True, color="red", alpha = 0.3)
# gdf_buurt_2023_fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="blue", alpha = 0.3)
#gdf_buurt_2023_fua[~gdf_buurt_2023_fua.fuaname_en.isna()].plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red")
#ax = gdf_20[gdf_20["BU_NAAM"] == "Verspreide huizen Sleen"].plot(figsize=(10, 10), edgecolor='k', linewidth=2.5)
# gdf_22[gdf_22["BU_CODE"] == "BU01145109"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'red')
#gdf_22[gdf_22["BU_NAAM"] == "Verspreide huizen Sleen"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'green')

# Set the title and display the plot
plt.title(f"Plot of NL", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")

# plt.xlim([min_longitude, max_longitude])  # Replace with your desired x-range (longitude)
# plt.ylim([min_latitude, max_latitude])    # Replace with your desired y-range (latitude)

plt.grid(True)
plt.show()

print(len(gdf_wijk_2023_fua.index.unique()))
print(len(gdf_wijk_2023_fua.index))
print(len(gdf_wijk_2023.index.unique()))
print(len(gdf_wijk_2023.index))
gdf_wijk_2023_fua[gdf_wijk_2023_fua.index.duplicated()]

gdf_wijk_2023_fua[gdf_wijk_2023_fua.WK_CODE.isin(["WK003459"])]

print(fua_buurt_2023.shape)
print(fua.shape)
print(gdf_wijk_2023.shape)
print(gdf_wijk_2023_fua.shape)

gdf_wijk_2023_fua.head()

# Quick plot
ax = gdf_buurt_2023.plot(figsize=(10, 10), edgecolor='k', linewidth=0.01, legend=True, color = "white")
# gdf_gemeente_2023_fua[~gdf_gemeente_2023_fua.fuaname_en.isna()].plot(ax = ax, column = 'NA_perc', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="Reds")
#fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red", alpha = 0.3)
gdf_wijk_2023_fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red", alpha = 0.3)
gdf_buurt_2023_fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="blue", alpha = 0.3)
#gdf_buurt_2023_fua[~gdf_buurt_2023_fua.fuaname_en.isna()].plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red")
#ax = gdf_20[gdf_20["BU_NAAM"] == "Verspreide huizen Sleen"].plot(figsize=(10, 10), edgecolor='k', linewidth=2.5)
# gdf_22[gdf_22["BU_CODE"] == "BU01145109"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'red')
#gdf_22[gdf_22["BU_NAAM"] == "Verspreide huizen Sleen"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'green')

# Set the title and display the plot
plt.title(f"Plot of NL", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")

# plt.xlim([min_longitude, max_longitude])  # Replace with your desired x-range (longitude)
# plt.ylim([min_latitude, max_latitude])    # Replace with your desired y-range (latitude)

plt.grid(True)
plt.show()

# Quick plot
ax = gdf_buurt_2023.plot(figsize=(10, 10), edgecolor='k', linewidth=0.01, legend=True, color = "white")
# gdf_gemeente_2023_fua[~gdf_gemeente_2023_fua.fuaname_en.isna()].plot(ax = ax, column = 'NA_perc', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="Reds")
#fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red", alpha = 0.3)
gdf_wijk_2023_fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red", alpha = 0.3)
gdf_buurt_2023_fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="blue", alpha = 0.3)
#gdf_buurt_2023_fua[~gdf_buurt_2023_fua.fuaname_en.isna()].plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red")
#ax = gdf_20[gdf_20["BU_NAAM"] == "Verspreide huizen Sleen"].plot(figsize=(10, 10), edgecolor='k', linewidth=2.5)
# gdf_22[gdf_22["BU_CODE"] == "BU01145109"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'red')
#gdf_22[gdf_22["BU_NAAM"] == "Verspreide huizen Sleen"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'green')

# Set the title and display the plot
plt.title(f"Plot of NL", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")

# plt.xlim([min_longitude, max_longitude])  # Replace with your desired x-range (longitude)
# plt.ylim([min_latitude, max_latitude])    # Replace with your desired y-range (latitude)

plt.grid(True)
plt.show()

gdf_buurt_2023.STED.unique()

# Quick plot
ax = gdf_wijk_2023.plot(figsize=(10, 10), edgecolor='k', linewidth=0.01, legend=True, color = "white")
# gdf_gemeente_2023_fua[~gdf_gemeente_2023_fua.fuaname_en.isna()].plot(ax = ax, column = 'NA_perc', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="Reds")
#fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red", alpha = 0.3)
gdf_wijk_2023[gdf_wijk_2023["STED"].isin([1,2])].plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.001, legend=True, color="blue", alpha = 0.3)
gdf_wijk_2023_fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.001, legend=True, color="red", alpha = 0.3)
# gdf_buurt_2023_fua.plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="blue", alpha = 0.3)
#gdf_buurt_2023_fua[~gdf_buurt_2023_fua.fuaname_en.isna()].plot(ax = ax,  figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, color="red")
#ax = gdf_20[gdf_20["BU_NAAM"] == "Verspreide huizen Sleen"].plot(figsize=(10, 10), edgecolor='k', linewidth=2.5)
# gdf_22[gdf_22["BU_CODE"] == "BU01145109"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'red')
#gdf_22[gdf_22["BU_NAAM"] == "Verspreide huizen Sleen"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'green')

# Set the title and display the plot
plt.title(f"Plot of NL", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")

# plt.xlim([min_longitude, max_longitude])  # Replace with your desired x-range (longitude)
# plt.ylim([min_latitude, max_latitude])    # Replace with your desired y-range (latitude)

plt.grid(True)
plt.show()

# Quick plot
ax = gdf_22.plot(column = ' LIHE_AndOR _LILEK', figsize=(10, 10), edgecolor='k', linewidth=0.5, legend=True, cmap="plasma")
#ax = gdf_20[gdf_20["BU_NAAM"] == "Verspreide huizen Sleen"].plot(figsize=(10, 10), edgecolor='k', linewidth=2.5)
# gdf_22[gdf_22["BU_CODE"] == "BU01145109"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'red')
#gdf_22[gdf_22["BU_NAAM"] == "Verspreide huizen Sleen"].plot(ax= ax, figsize=(10, 10), edgecolor='k', linewidth=0.5, color = 'green')

# Set the title and display the plot
plt.title(f"Plot of NL", fontsize=15)
plt.xlabel("Longitude")
plt.ylabel("Latitude")
plt.grid(True)
plt.show()

energy_poverty_22.head()

gdf_22[~gdf_22.GM_NAAM.isin(energy_poverty_22.Gemeente)]["GM_NAAM"]

gdf_22[gdf_22['BU_NAAM'] == "Zaagpoortbuurt"]

gdf_21[gdf_21['BU_NAAM'] == "Zaagpoortbuurt"]

gdf_20[gdf_20['BU_NAAM'] == "Zaagpoortbuurt"]

gdf_19[gdf_19['BU_NAAM'] == "Zaagpoortbuurt"]
