==> Python/01_data_processing_geometries_19_23.ipynb <==
import pandas as pd 
import os
import geopandas as gpd 
import zipfile 
import matplotlib.pyplot as plt
import folium
import pickle 
import numpy as np 

os.chdir("C:/Users/tejed002/OneDrive - Wageningen University & Research/PhD-WUR/01_WUR-Research/05_Papers/02_Paper_energy_poverty/Data/Raw")

# with zipfile.ZipFile('ShapeFiles/WijkBuurtkaart_2019_v3.zip', 'r') as zip_ref:
#        zip_ref.extractall("ShapeFiles/WijkBuurtkaart_2019")
# with zipfile.ZipFile('ShapeFiles/WijkBuurtkaart_2020_v3.zip', 'r') as zip_ref:
#        zip_ref.extractall("ShapeFiles/WijkBuurtkaart_2020")
# with zipfile.ZipFile('ShapeFiles/WijkBuurtkaart_2021_v3.zip', 'r') as zip_ref:
#     zip_ref.extractall("ShapeFiles/WijkBuurtkaart_2021")
# with zipfile.ZipFile('ShapeFiles/WijkBuurtkaart_2022_v2.zip', 'r') as zip_ref:
#        zip_ref.extractall("ShapeFiles/WijkBuurtkaart_2022")

buurt_traceable = {}
for year in ["2021_2022", "2020_2021", "2019_2020"]:
    buurt_traceable[year] = pd.read_excel("C:/Users/tejed002/OneDrive - Wageningen University & Research/PhD-WUR/01_WUR-Research/05_Papers/02_Paper_energy_poverty/Data/Raw/buurten_consistency_table_2019_2022.xlsx", sheet_name = year)                                   

dict_wijk_gdf = {}
dict_gdf = {}
dict_gemeente_gdf = {}
years = [2019, 2020, 2021, 2022, 2023]

for year in years:
    if year in [2019, 2020, 2021]:
        version = 3
    elif year == 2022:
        version = 2
    else:
        version = 1
    if year in [2021, 2022, 2023]:
        file_name = f"ShapeFiles/WijkBuurtkaart_{year}/wijken_{year}_v{version}.shp"
    else:
        file_name = f"ShapeFiles/WijkBuurtkaart_{year}/wijk_{year}_v{version}.shp" 
    print(file_name)
    dict_wijk_gdf[str(year)] = gpd.read_file(file_name)

dict_wijk_gdf['2022'].shape

for year in years:
    if year in [2019, 2020, 2021]:
        version = 3
    elif year == 2022:
        version = 2
    else:
        version = 1
    if year in [2021, 2022, 2023]:
        file_name = f"ShapeFiles/WijkBuurtkaart_{year}/gemeenten_{year}_v{version}.shp"
    else:
        file_name = f"ShapeFiles/WijkBuurtkaart_{year}/gemeente_{year}_v{version}.shp" 
    print(file_name)
    dict_gemeente_gdf[str(year)] = gpd.read_file(file_name)

for year in years:
    if year in [2019, 2020, 2021]:
        version = 3
    elif year == 2022:
        version = 3
    else:
        version == 2
    file_name = f"ShapeFiles/WijkBuurtkaart_{year}/buurten_{year}_v{version}.shp"
    dict_gdf[str(year)] = gpd.read_file(file_name)

# Consolidate 2022 as the final Energy Poverty Monitor. 
# # Standardize all the neighborhoods for 2022, and bring all the other years to 2022
# # Use the SHP file to do that, and the neighborhood change Excel file 
# # 


with open("ShapeFiles/00_Processed/dict_gdf.pkl", "wb") as file:
    pickle.dump(dict_gdf, file)

with open("ShapeFiles/00_Processed/dict_wijk_gdf.pkl", "wb") as file:
    pickle.dump(dict_wijk_gdf, file)

with open("ShapeFiles/00_Processed/dict_gemeente_gdf.pkl", "wb") as file:
    pickle.dump(dict_gemeente_gdf, file)

years = [2019, 2020, 2021, 2022, 2023]
for year in years:
    gdf = dict_gdf[str(year)]
    gdf = gdf[gdf["H2O"] == "NEE"]
    dict_gdf[str(year)] = gdf

columns_buurten = ["BU_CODE", "BU_NAAM", "WK_CODE", "GM_CODE", "GM_NAAM", "WK_CODE", "GM_CODE", "GM_NAAM", "IND_WBI", 'JRSTATCODE', 'JAAR', 'geometry']

gdf_22 = dict_gdf["2022"][columns_buurten] 
gdf_21 = dict_gdf["2021"][columns_buurten] 
gdf_20 = dict_gdf["2020"][columns_buurten] 
gdf_19 = dict_gdf["2019"][columns_buurten] 
buurt_2019_2022 = gdf_22[columns_buurten] 
buurt_2019_2022["year"] = 2022

#### checking if homologation table has only unique change of codes
print(buurt_traceable["2021_2022"].BU_WK_CODE_2021.nunique())
print(buurt_traceable["2021_2022"].BU_WK_CODE_2022.nunique())

print(buurt_traceable["2020_2021"].BU_WK_CODE_2020.nunique())
print(buurt_traceable["2020_2021"].BU_WK_CODE_2021.nunique())

print(buurt_traceable["2019_2020"].BU_WK_CODE_2019.nunique())
print(buurt_traceable["2019_2020"].BU_WK_CODE_2020.nunique())


BU_CODE_2019 = gdf_19[["BU_CODE"]]
BU_CODE_2019 = BU_CODE_2019.rename(columns = {"BU_CODE" : "BU_CODE_2019"})
BU_CODE_2019.head()

BU_CODE_2020 = gdf_20[["BU_CODE"]]
BU_CODE_2020 = BU_CODE_2020.rename(columns = {"BU_CODE" : "BU_CODE_2020"})
BU_CODE_2020.head()

BU_CODE_2021 = gdf_21[["BU_CODE"]]
BU_CODE_2021 = BU_CODE_2021.rename(columns = {"BU_CODE" : "BU_CODE_2021"})
BU_CODE_2021.head()

### Homologate neighborhoods that is possible with the Excel file. Then missing ones can be done with spatial join (sjoin)
### 2021 --> 2022 
buurt_2019_2022 =  pd.merge(buurt_2019_2022, BU_CODE_2021, left_on = "BU_CODE", right_on = "BU_CODE_2021", how = "left", indicator=True)

# BU_CODE_2021.BU_CODE_2021.isin(buurt_traceable["2021_2022"].BU_WK_CODE_2021).value_counts()
buurt_traceable["2021_2022"].BU_WK_CODE_2021.isin(BU_CODE_2021.BU_CODE_2021).value_counts()


print(buurt_2019_2022["BU_CODE_2021"].isna().value_counts())
buurt_2019_2022 = pd.merge(buurt_2019_2022, buurt_traceable["2021_2022"][["BU_WK_CODE_2021", "BU_WK_CODE_2022"]], left_on = "BU_CODE", right_on = "BU_WK_CODE_2022", how = "left")
buurt_2019_2022["BU_CODE_2021"] = np.where(buurt_2019_2022["BU_CODE_2021"].isna(), buurt_2019_2022["BU_WK_CODE_2021"], buurt_2019_2022["BU_CODE_2021"]) 
print(buurt_2019_2022["BU_CODE_2021"].isna().value_counts())
# miss21= buurt_2019_2022[~buurt_2019_2022.BU_CODE.isin(gdf_21.BU_CODE_2022)]
# miss21.shape

buurt_2019_2022 =  buurt_2019_2022.drop(columns = ["_merge", "BU_WK_CODE_2021", "BU_WK_CODE_2022"])
buurt_2019_2022.head()

### 2020 --> 2022 
buurt_2019_2022 =  pd.merge(buurt_2019_2022, BU_CODE_2020, left_on = "BU_CODE", right_on = "BU_CODE_2020", how = "left", indicator=True)
buurt_2019_2022.head()

buurt_2019_2022['_merge'].value_counts()

# BU_CODE_2021.BU_CODE_2021.isin(buurt_traceable["2021_2022"].BU_WK_CODE_2021).value_counts()
print(buurt_traceable["2020_2021"].BU_WK_CODE_2020.isin(buurt_2019_2022[buurt_2019_2022['_merge'] == "left_only"].BU_CODE).value_counts())
print(buurt_traceable["2020_2021"].BU_WK_CODE_2021.isin(buurt_2019_2022[buurt_2019_2022['_merge'] == "left_only"].BU_CODE).value_counts())


print(buurt_2019_2022["BU_CODE_2020"].isna().value_counts())
buurt_2019_2022 = pd.merge(buurt_2019_2022, buurt_traceable["2020_2021"][["BU_WK_CODE_2020", "BU_WK_CODE_2021"]], left_on = "BU_CODE", right_on = "BU_WK_CODE_2021", how = "left")
buurt_2019_2022["BU_CODE_2020"] = np.where(buurt_2019_2022["BU_CODE_2020"].isna(), buurt_2019_2022["BU_WK_CODE_2020"], buurt_2019_2022["BU_CODE_2020"]) 
print(buurt_2019_2022["BU_CODE_2020"].isna().value_counts())


buurt_2019_2022 = buurt_2019_2022.drop(columns = ["BU_WK_CODE_2020", "BU_WK_CODE_2021"]) 

print(buurt_traceable["2021_2022"].BU_WK_CODE_2022.isin(buurt_2019_2022[buurt_2019_2022["BU_CODE_2020"].isna()].BU_CODE).value_counts())
print(buurt_traceable["2021_2022"].BU_WK_CODE_2021.isin(buurt_2019_2022[buurt_2019_2022["BU_CODE_2020"].isna()].BU_CODE).value_counts())

print(buurt_2019_2022["BU_CODE_2020"].isna().value_counts())
buurt_2019_2022 = pd.merge(buurt_2019_2022, buurt_traceable["2021_2022"][["BU_WK_CODE_2022", "BU_WK_CODE_2021"]], left_on = "BU_CODE", right_on = "BU_WK_CODE_2022", how = "left")
buurt_2019_2022["BU_CODE_2020"] = np.where(buurt_2019_2022["BU_CODE_2020"].isna(), buurt_2019_2022["BU_WK_CODE_2021"], buurt_2019_2022["BU_CODE_2020"]) 
print(buurt_2019_2022["BU_CODE_2020"].isna().value_counts())

print(buurt_2019_2022["BU_CODE_2020"].isin(gdf_20["BU_CODE"]).value_counts())
### Delete inconsistencies 
buurt_2019_2022["BU_CODE_2020"] = np.where(~buurt_2019_2022["BU_CODE_2020"].isin(gdf_20["BU_CODE"]), pd.NA, buurt_2019_2022["BU_CODE_2020"]) 
print(buurt_2019_2022["BU_CODE_2020"].isin(gdf_20["BU_CODE"]).value_counts())

buurt_2019_2022 = buurt_2019_2022.drop(columns = ["_merge", "BU_WK_CODE_2022", "BU_WK_CODE_2021"])

### 2019 -- 2022 
buurt_2019_2022 =  pd.merge(buurt_2019_2022, BU_CODE_2019, left_on = "BU_CODE", right_on = "BU_CODE_2019", how = "left", indicator=True)
buurt_2019_2022.head()

# BU_CODE_2021.BU_CODE_2021.isin(buurt_traceable["2021_2022"].BU_WK_CODE_2021).value_counts()
print(buurt_traceable["2019_2020"].BU_WK_CODE_2019.isin(buurt_2019_2022[buurt_2019_2022['_merge'] == "left_only"].BU_CODE).value_counts())
print(buurt_traceable["2019_2020"].BU_WK_CODE_2020.isin(buurt_2019_2022[buurt_2019_2022['_merge'] == "left_only"].BU_CODE).value_counts())

print(buurt_2019_2022["BU_CODE_2019"].isna().value_counts())
buurt_2019_2022 = pd.merge(buurt_2019_2022, buurt_traceable["2019_2020"][["BU_WK_CODE_2019", "BU_WK_CODE_2020"]], left_on = "BU_CODE", right_on = "BU_WK_CODE_2020", how = "left")
buurt_2019_2022["BU_CODE_2019"] = np.where(buurt_2019_2022["BU_CODE_2019"].isna(), buurt_2019_2022["BU_WK_CODE_2019"], buurt_2019_2022["BU_CODE_2019"]) 
print(buurt_2019_2022["BU_CODE_2019"].isna().value_counts())


buurt_2019_2022 = buurt_2019_2022.drop(columns = ["BU_WK_CODE_2019", "BU_WK_CODE_2020"]) 

print(buurt_traceable["2021_2022"].BU_WK_CODE_2022.isin(buurt_2019_2022[buurt_2019_2022["BU_CODE_2019"].isna()].BU_CODE).value_counts())
print(buurt_traceable["2021_2022"].BU_WK_CODE_2021.isin(buurt_2019_2022[buurt_2019_2022["BU_CODE_2019"].isna()].BU_CODE).value_counts())

print(buurt_2019_2022["BU_CODE_2019"].isna().value_counts())
buurt_2019_2022 = pd.merge(buurt_2019_2022, buurt_traceable["2021_2022"][["BU_WK_CODE_2022", "BU_WK_CODE_2021"]], left_on = "BU_CODE", right_on = "BU_WK_CODE_2022", how = "left")
buurt_2019_2022["BU_CODE_2019"] = np.where(buurt_2019_2022["BU_CODE_2019"].isna(), buurt_2019_2022["BU_WK_CODE_2021"], buurt_2019_2022["BU_CODE_2019"]) 
print(buurt_2019_2022["BU_CODE_2019"].isna().value_counts())

print(buurt_2019_2022["BU_CODE_2019"].isin(gdf_19["BU_CODE"]).value_counts())
### Delete inconsistencies 
buurt_2019_2022["BU_CODE_2019"] = np.where(~buurt_2019_2022["BU_CODE_2019"].isin(gdf_19["BU_CODE"]), pd.NA, buurt_2019_2022["BU_CODE_2019"]) 
print(buurt_2019_2022["BU_CODE_2019"].isin(gdf_19["BU_CODE"]).value_counts())

buurt_2019_2022 =  buurt_2019_2022.drop(columns = ["_merge", "BU_WK_CODE_2022", "BU_WK_CODE_2021"])
buurt_2019_2022.head()

# Export temporal file 
outpath =  "C:/Users/tejed002/OneDrive - Wageningen University & Research/PhD-WUR/01_WUR-Research/05_Papers/02_Paper_energy_poverty/Data/Raw"
buurt_2019_2022.to_csv(os.path.join(outpath, "buurt_homologation_2019_2022.csv"), sep = '|')
